[LIT] Fix system-windows
Turns out if you were on windows and your default target wasn't windows the system-windows feature wasn't getting enabled.

This fixes that and updates the coff-dwarf test to rely on the new "target-windows" feature. That test was the reason why system-windows was changed to not always be enabled on Windows hosts.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@289503 91177308-0d34-0410-b5e6-96231b3b80d8